In the attestion protocol, encryption using simetric keys is now configurable via xml config file and now during DH the server sends the configuration to the client.

+++ SRSC-P2/src/main/java/GetSetBenchmark.java
@@ -4,35 +4,45 @@ import java.io.IOException;
 import java.net.UnknownHostException;
 import java.util.Calendar;
 
+import TPM.client.TPMClient;
 import redis.clients.jedis.Jedis;
 
 public class GetSetBenchmark {
-    private static final int TOTAL_OPERATIONS = 100;
+	private static final int TOTAL_OPERATIONS = 100;
 
-    public static void main(String[] args) throws UnknownHostException, IOException {
-        //Jedis jedis = new Jedis("rediss://localhost", 6379);
-        Jedis jedis = new Jedis("172.17.0.2", 6379);
+	public static void main(String[] args) throws UnknownHostException, IOException {
+		
+		//if attestion ok do benchmark
+		if(	new TPMClient().attest("localhost", 4446, "localhost", 4443,
+				"TPMClientTrustStore", "srscsrsc", "gostpmservercert", "vmstpmservercert")) {
+			
+			
+			//Jedis jedis = new Jedis("rediss://localhost", 6379);
+			Jedis jedis = new Jedis("172.17.0.2", 6379);
 
-        //jedis.connect();
-        jedis.flushAll();
+			//jedis.connect();
+			jedis.flushAll();
 
-        long begin = Calendar.getInstance().getTimeInMillis();
+			long begin = Calendar.getInstance().getTimeInMillis();
 
-        for (int n = 0; n <= TOTAL_OPERATIONS; n++) {
-            String key = "foo" + n;
-            jedis.set(key, "bar" + n);
-            // System.out.println(jedis.get(key));
-        }
-        for (int n = 0; n <= TOTAL_OPERATIONS; n++) {
-            String key = "foo" + n;
-            jedis.get(key);
-            //System.out.println(jedis.get(key));
-        }
+			for (int n = 0; n <= TOTAL_OPERATIONS; n++) {
+				String key = "foo" + n;
+				jedis.set(key, "bar" + n);
+				// System.out.println(jedis.get(key));
+			}
+			for (int n = 0; n <= TOTAL_OPERATIONS; n++) {
+				String key = "foo" + n;
+				jedis.get(key);
+				//System.out.println(jedis.get(key));
+			}
 
-        long elapsed = Calendar.getInstance().getTimeInMillis() - begin;
+			long elapsed = Calendar.getInstance().getTimeInMillis() - begin;
 
-        jedis.disconnect();
+			jedis.disconnect();
 
-        System.out.println(((1000 * 2 * TOTAL_OPERATIONS) / elapsed) + " ops");
-    }
+			System.out.println(((1000 * 2 * TOTAL_OPERATIONS) / elapsed) + " ops");	
+		}
+
+
+	}
 }
